Move classlist handlers into own file
This also involves moving the DOM element references into their own
module.
